Moved UI components to UI subpackage
