+++ ui/src/components/BankScale.js
